feat: update tournament

+++ src/lib/utils.ts
@@ -88,7 +88,7 @@ export const formatDateTime = (dateString: Date | string | null | undefined) => 
   };
   const dateOptions: Intl.DateTimeFormatOptions = {
     weekday: 'short', // abbreviated weekday name (e.g., 'Mon')
-    month: 'short', // abbreviated month name (e.g., 'Oct')
+    month: 'numeric', // abbreviated month name (e.g., 'Oct')
     year: 'numeric', // abbreviated year (e.g., '2023')
     day: 'numeric', // numeric day of the month (e.g., '25')
   };
